get app name synced up, fix several bugs
