+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.curious-app",
+    package = "com.curiousapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.curious-app",
+    package = "com.curiousapp",
     res = "src/main/res",
 )
