+++ android/app/build.gradle
@@ -98,7 +98,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.curious-app"
+        applicationId "com.curiousapp"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
