+++ android/app/src/main/java/com/curiousapp/MainActivity.java
@@ -1,17 +1,8 @@
 package com.curiousapp;
 
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
-import android.content.Intent;
-import com.dieam.reactnativepushnotification.ReactNativePushNotificationPackage;
 
 public class MainActivity extends ReactActivity {
-
-    private ReactNativePushNotificationPackage mReactNativePushNotificationPackage;
 
     /**
      * Returns the name of the main component registered from JavaScript.
@@ -19,35 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "curiousapp";
+        return "Curious";
     }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    protected List<ReactPackage> getPackages() {
-        mReactNativePushNotificationPackage = new ReactNativePushNotificationPackage(this);
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            mReactNativePushNotificationPackage
-        );
-    }
-
-    // Add onNewIntent
-    @Override
-    public void onNewIntent (Intent intent) {
-      super.onNewIntent(intent);
-
-      mReactNativePushNotificationPackage.newIntent(intent);
-    }
-
 }
