+++ android/settings.gradle
@@ -1,3 +1,5 @@
-rootProject.name = 'curious-app'
+rootProject.name = 'curiousapp'
 
 include ':app'
+include ':react-native-push-notification'
+project(':react-native-push-notification').projectDir = file('../node_modules/react-native-push-notification/android')
