+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "curious-app",
-  "displayName": "curious-app"
+  "name": "curiousapp",
+  "displayName": "Curious"
 }
