+++ app/index.js
@@ -1,25 +1,75 @@
-import React from 'react';
-import { StyleSheet, View, Text } from 'react-native';
-import Welcome from './src/components/Welcome';
-
-export default class App extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text>Open up App.js to start working on your app!</Text>
-        <Text>Changes you make will automatically reload.</Text>
-        <Text>Shake your phone to open the developer menu.</Text>
-        <Welcome />
-      </View>
-    );
-  }
-}
+import React, { Component } from 'react';
+import { View, Text, StyleSheet, Picker, AppState, Platform } from 'react-native';
+import PushController from './src/components/PushController';
+import PushNotification from 'react-native-push-notification';
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#fff',
+    justifyContent: 'center',
     alignItems: 'center',
-    justifyContent: 'center'
-  }
+    backgroundColor: '#F5FCFF',
+  },
+  welcome: {
+    fontSize: 20,
+    textAlign: 'center',
+    margin: 10,
+  },
+  picker: {
+    width: 100,
+  },
 });
+
+export default class App extends Component {
+  constructor(props) {
+    super(props);
+
+    this.handleAppStateChange = this.handleAppStateChange.bind(this);
+    this.state = {
+      seconds: 5,
+    };
+  }
+
+  componentDidMount() {
+    AppState.addEventListener('change', this.handleAppStateChange);
+  }
+
+  componentWillUnmount() {
+    AppState.removeEventListener('change', this.handleAppStateChange);
+  }
+
+  handleAppStateChange(appState) {
+    if (appState === 'background') {
+      let date = new Date(Date.now() + (this.state.seconds * 1000));
+
+      if (Platform.OS === 'ios') {
+        date = date.toISOString();
+      }
+
+      PushNotification.localNotificationSchedule({
+        message: "My Notification Message",
+        date,
+      });
+    }
+  }
+
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.welcome}>
+          Choose your notification time in seconds.
+        </Text>
+        <Picker
+          style={styles.picker}
+          selectedValue={this.state.seconds}
+          onValueChange={(seconds) => this.setState({ seconds })}
+        >
+          <Picker.Item label="5" value={5} />
+          <Picker.Item label="10" value={10} />
+          <Picker.Item label="15" value={15} />
+        </Picker>
+        <PushController />
+      </View>
+    );
+  }
+}
