+++ index.android.js
@@ -8,4 +8,4 @@ import React from 'react';
 import { AppRegistry } from 'react-native';
 import App from './app';
 
-AppRegistry.registerComponent('PushNotificationHowTo', () => App);
+AppRegistry.registerComponent('Curious', () => App);
