--- index.ios.js
@@ -1,11 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
-
-import React from 'react';
-import { AppRegistry } from 'react-native';
-import App from './app';
-
-AppRegistry.registerComponent('PushNotificationHowTo', () => App);
