+++ package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "react": "16.0.0-beta.5",
     "react-native": "^0.49.5",
-    "react-native-push-notification": "^1.0.7"
+    "react-native-push-notification": "^1.0.7",
+    "stacktrace-parser": "^0.1.4"
   }
 }
